+++ site/.vuepress/components/codes/codegen.js
@@ -1,7 +1,9 @@
 import InstantRestart from './InstantRestart/codegen.js';
+import qft from './qft/codegen.js';
 
 export default {
   InstantRestart,
+  qft,
 };
 
 /**